requirements.txt
asgiref==3.2.7
astroid==2.2.5
async-timeout==3.0.1
attrs==19.1.0
certifi==2019.9.11
chardet==3.0.4
colorama==0.4.1
dj-database-url==0.5.0
Django==3.0.5
django-crispy-forms==1.7.2
django-heroku==0.3.1
djangorestframework==3.11.0
gunicorn==19.9.0
h11==0.8.1
h2==3.1.1
hpack==3.0.0
http3==0.6.7
hyperframe==5.2.0
idna==2.8
isort==4.3.21
lazy-object-proxy==1.4.2
mccabe==0.6.1
multidict==4.5.2
psycopg2==2.8.5
psycopg2-binary==2.8.5
pylint==2.3.1
pylint-django==2.0.11
pylint-plugin-utils==0.5
pytz==2019.2
requests==2.22.0
rfc3986==1.3.2
scikit-learn==0.21.1
six==1.12.0
sqlparse==0.3.1
typed-ast==1.4.0
urllib3==1.25.3
whitenoise==4.1.3
wrapt==1.11.2
yarl==1.3.0
pandas==0.24.2
pickleshare==0.7.5
